wip(core): move files from store to the 'core' directory
