+++ src/core/compose.ts
